% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_missing.R
\name{convert_missing}
\alias{convert_missing}
\title{Clean up missing values in Wisconsin WRAP/ADRC data}
\usage{
convert_missing(x, raw.name, source)
}
\arguments{
\item{x}{Vector containing untransformed Wisconsin WRAP/ADRC data.}

\item{raw.name}{Name of the original variable. Should be among the following: anartraw, ANIMALS, AnimTotRaw, bnttot, bnttot30, BOSTON, CRAFTDRE, CRAFTURS, drraw, flucfl, LOGIMEM, MEMUNITS, MINTTOTS, ORRT.Unc_Std, readstn, TRAILA, TRAILB, trla, trlb, tTotal, ttotal, UDSVERTN, waisrtot, wmsrar, wmsrar2}

\item{source}{Data source. Should be one of the following: ADRC, WRAP}
}
\value{
Vector containing cleaned Wisconsin WRAP/ADRC data.
}
\description{
This function takes in a column of raw Wisconsin WRAP/ADRC data and returns a column with missing value codes replaced as NA.
}
\seealso{
\code{\link{wis_harm}}
}
\keyword{ADRC}
\keyword{WRAP,}
\keyword{Wisconsin,}
