% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_values.R
\name{recode_values}
\alias{recode_values}
\title{Recode a single Wisconsin WRAP/ADRC data value}
\usage{
recode_values(x, raw.name, source = Data.Source)
}
\arguments{
\item{x}{Vector containing untransformed UP data.}

\item{raw.name}{Name of the original variable. Should be among the following: anartraw, ANIMALS, AnimTotRaw, bnttot, bnttot30, BOSTON, CRAFTDRE, CRAFTURS, drraw, flucfl, LOGIMEM, MEMUNITS, MINTTOTS, ORRT.Unc_Std, readstn, TRAILA, TRAILB, trla, trlb, tTotal, ttotal, UDSVERTN, waisrtot, wmsrar, wmsrar2}

\item{source}{Data source. Should be one of the following: ADRC, WRAP}
}
\value{
Vector containing recoded Wisconsin WRAP/ADRC data.
}
\description{
This function takes in a column of raw Wisconsin WRAP/ADRC data and returns a column of crosswalked Wisconsin WRAP/ADRC data. It is intended as an internal function.
}
\seealso{
\code{\link{wis_harm}}
}
\keyword{ADRC}
\keyword{WRAP,}
\keyword{Wisconsin,}
