% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis_composites.R
\name{wis_composites}
\alias{wis_composites}
\title{Compute crosswalked composite scores in Wisconsin WRAP/ADRC}
\usage{
wis_composites(data, id, source, visno)
}
\arguments{
\item{data}{Vector containing untransformed Wisconsin WRAP/ADRC data.}

\item{id}{Column containing subject ID. Defaults to Reggieid.}

\item{source}{Column containing data source. Defaults to Data.Source.}

\item{visno}{Column containing visit number. Defaults to Visit_Number.}
}
\value{
Data frame containing Wisconsin WRAP/ADRC data augmented with composites.
}
\description{
This function takes in a column of raw Wisconsin WRAP/ADRC data and returns the crosswalked version
}
\seealso{
\code{\link{wis_est}, \link{wis_harm}}
}
\keyword{ADRC,}
\keyword{WRAP,}
\keyword{Wisconsin,}
\keyword{crosswalk,}
\keyword{harmonization}
