% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis_est.R
\name{wis_est}
\alias{wis_est}
\title{Estimate unconditional and conditional percentiles using an rrq() object}
\usage{
wis_est(
  data,
  variable.list = NULL,
  model.unc = NULL,
  model.c = NULL,
  raw = TRUE,
  raw.names = NULL,
  id = "subjid",
  source = "Data.Source",
  age = "age",
  visno = "Visit_Number",
  gender = "gender",
  ed = "Education.Years"
)
}
\arguments{
\item{data}{Data frame containing neuropsych scores and covariates.}

\item{variable.list}{Vector containing names of variables to evaluate. Norms are available for the following (crosswalked) outcomes: animtotraw, bnt.xw, cfl.xw, drraw, dspf.xw, dspb.xw, lm_del.xw, lm_imm.xw, PACC3.cfl.xw.sca, PACC3.trlb.xw.sca, PACC4.cfl.trlb.xw.sca, PACC4.trlb.xw.sca, PACC3.wrap.sca, PACC4.wrap.sca, PACC5.wrap.sca, theo.mem.xw.sca, trla, trlb.xw, ttotal.}

\item{model.unc}{Quantile regression model object used to predict unconditional percentile. Default model included in package built using Freeze 17 data.}

\item{model.c}{Quantile regression model object used to predict conditional percentile. Default model included in package built using Freeze 17 data.}

\item{raw}{Boolean variable indicating whether the data are raw data. If set to TRUE (the default), the values in \code{raw.name} will be crosswalked to the variable specified in \code{variable} using crosswalks developed by Rebecca Langhough Koscik.}

\item{raw.names}{Name of columns in \code{data} containing original data values to be transformed to \code{variable}.}

\item{id}{Name of column containing patient ID.}

\item{source}{Name of column containing data source (WRAP or ADRC).}

\item{age}{Name of column containing age.}

\item{visno}{Name of column containing visit number.}

\item{gender}{Name of column containing gender.}

\item{ed}{Name of column containing years of education.}
}
\value{
Dataframe containing values best.q.unc and best.q.c corresponding to the best-fit unconditional and conditional percentiles, respectively.
}
\description{
This function takes in a neuropsych dataset and a quantile regression object and computes the best-fit unconditional and conditional percentiles for each observation against a reference dataset. By default, the quantile regression object is a model of the specified variable based on Wisconsin WRAP and ADRC data from Freeze 17. The results are returned as a data frame.
}
\seealso{
\code{\link{wis_harm}}
}
\keyword{ADRC,}
\keyword{WRAP,}
\keyword{Wisconsin,}
\keyword{conditional}
\keyword{norms}
