% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis_harm.R
\name{wis_harm}
\alias{wis_harm}
\title{Convert raw to crosswalked scores in Wisconsin WRAP/ADRC}
\usage{
wis_harm(
  data,
  raw.names,
  id = "Reggieid",
  source = "Data.Source",
  visno = "Visit_Number"
)
}
\arguments{
\item{data}{Vector containing untransformed Wisconsin WRAP/ADRC data.}

\item{raw.names}{Names of the original variables. Should be among the following: anartraw, ANIMALS, AnimTotRaw, bnttot, bnttot30, BOSTON, CRAFTDRE, CRAFTURS, drraw, flucfl, LOGIMEM, MEMUNITS, MINTTOTS, ORRT.Unc_Std, readstn, TRAILA, TRAILB, trla, trlb, tTotal, ttotal, UDSVERTN, waisrtot, wmsrar, wmsrar2}

\item{id}{Column containing subject ID. Defaults to Reggieid.}

\item{source}{Column containing data source. Defaults to Data.Source.}

\item{visno}{Column containing visit number. Defaults to Visit_Number.}
}
\value{
Data frame containing crosswalked Wisconsin WRAP/ADRC data.
}
\description{
This function takes in a column of raw Wisconsin WRAP/ADRC data and returns the crosswalked version
}
\seealso{
\code{\link{wis_est}}
}
\keyword{ADRC,}
\keyword{WRAP,}
\keyword{Wisconsin,}
\keyword{crosswalk,}
\keyword{harmonization}
