% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis_plot.R
\name{wis_plot}
\alias{wis_plot}
\title{Convert raw to crosswalked scores in Wisconsin WRAP/ADRC}
\usage{
wis_plot(
  data,
  var,
  sub,
  vislabel = FALSE,
  biomarkers = NULL,
  path = NULL,
  width = 15,
  height = 10
)
}
\arguments{
\item{data}{Vector containing the output of a wis_est call.}

\item{var}{Names of the original variables. Should be among the following: anartraw, ANIMALS, AnimTotRaw, bnttot, bnttot30, BOSTON, CRAFTDRE, CRAFTURS, drraw, flucfl, LOGIMEM, MEMUNITS, MINTTOTS, ORRT.Unc_Std, readstn, TRAILA, TRAILB, trla, trlb, tTotal, ttotal, UDSVERTN, waisrtot, wmsrar, wmsrar2}

\item{sub}{Column containing subject ID. Defaults to Reggieid.}

\item{vislabel}{Column containing data source. Defaults to Data.Source.}

\item{biomarkers}{Column containing visit number. Defaults to Visit_Number.}

\item{path}{File path in which to store output plot. The plot filename will be created based on the sub and var parameters.}

\item{width}{Width of output png file.}

\item{height}{Height of output png file.}
}
\value{
Plot containing indidvidual longitudinal trajectories for selected WisNorms variables.
}
\description{
This function takes in an output dataset matching the output of wis_est, and a variable or variable list and
subject id, and outputs a plot. This is a parallel version of the Shiny app in runExample() that can be used
at the console. If given a path, this function will output the plot as a .png file. Otherwise, it will output
the plot to the current session.
}
\seealso{
\code{\link{wis_est}}
}
\keyword{ADRC,}
\keyword{WRAP,}
\keyword{Wisconsin,}
\keyword{crosswalk,}
\keyword{harmonization}
